requirements.txt
discord.py[voice]
ksoftapi
jishaku
statcord.py
asyncpg
python-dateutil